requirements.txt
flask
flask-cors
requests
gunicorn
google-cloud-storage